Refactor _join_url
