+++ alibi_detect/models/tensorflow/__init__.py
@@ -1,12 +1,3 @@
-# ###### CHANGED in TF-TFP-TORCH OPTIONAL DEPS PR ####################
-# ###### these comments should be removed prior to PR merge. #########
-# ####################################################################
-#
-# 1. loss functions that require tensorflow-probability have been
-# split out into `prob_losses.py` in order to  fix import errors for
-# tensorflow_probability.
-
-
 from alibi_detect.utils.missing_optional_dependency import import_optional
